10 hours of work later
2 objects playable, stopped flickering, created my prefabs folder and
created two prefabs one for platform and one for player, player
movements in sync, platforms in sync BUT extra platform appears when
player 2 joins, camera no longer follows players (because players aren't
actually an object but a prefab and an instantiated object upon game
creation...nothing for camera to snap to pre compile), fix sync issues
with second platform.  Timer needs to be setup to start upon game
start...which requires both players to be present.